Crate bchannel

Structs

Receiver
A better Receiver struct. A receiver can receive values of type T and can contain at most one error of type E.

ReceiverIterator
An iterator wrapper for receivers.

Sender
A better Sender struct. A sender can send values of type T and can signal an error of type E. (It can only signal one error).

Functions

channel
Creates a linked channel between a Sender and Receiver.