Struct bchannel::ReceiverIterator

pub struct ReceiverIterator<'a, T: 'a, E: 'a> {
    // some fields omitted
}

An iterator over received items.

This struct can either own or have a reference to the receiver that it gets its elements from.

This struct can either block when waiting for a message, or it can finish early (and be reusable) when it runs out of messages in the queue.